Moved QuatFromAngle etc to dmsdk/dlib/vmath.h
